Refactor quiz scoring for dynamic passing scores
Updated `QuizAttempt` to use `decimal?` for `PassingScore`, enabling precise thresholds. Modified `IQuizScoringService` to add a new `Score` overload with a `passingScore` parameter while maintaining backward compatibility. Refactored `QuizScoringService` to validate and use dynamic passing scores, replacing the hardcoded 70% threshold. Removed redundant methods.

Updated `QuizService` to pass `quiz.PassingScore` or default to 70.00m. Adjusted hint logic to align with the new AI response structure, using `AIHint` objects. Improved maintainability and ensured compatibility with updated data contracts.

+++ backend/src/MentraAI.API/Modules/Quizzes/Models/QuizAttempt.cs
@@ -20,7 +20,7 @@ public class QuizAttempt
     public DateTime? SubmittedAt { get; set; }
 
     // NEW — from AI response quiz.passing_score and quiz.time_limit_minutes
-    public int? PassingScore { get; set; }
+    public decimal? PassingScore { get; set; }
     public int? TimeLimitMinutes { get; set; }
 
     public UserStageProgress StageProgress { get; set; } = null!;
